Add support for expansion of $GEM_HOME and $GEM_PATH env variables.
Expansion of $GEM_HOME, ${GEM_HOME}, $GEM_PATH, and ${GEM_PATH} has been
implemented. This includes support for rvm and project-specific settings
(usually specified with .ruby-gemset and .ruby-version files); initial
expansion, upon Xcode startup, will be global (or the default gemset).
This is an Xcode limitation. When selecting a Cocoapods menu option, the
project path is determined via the key window (a hack, but there is no
other way). When $GEM_PATH is specified, path components are extracted
and applied individually for command execution.